+++ api/src/routes/mod.rs
@@ -24,7 +24,7 @@ impl ResponseError for UniversalInboxError {
             UniversalInboxError::InvalidUrlData { .. } => StatusCode::INTERNAL_SERVER_ERROR,
             UniversalInboxError::InvalidInputData { .. } => StatusCode::BAD_REQUEST,
             UniversalInboxError::InvalidParameters { .. } => StatusCode::BAD_REQUEST,
-            UniversalInboxError::TaskNotFound { .. } => StatusCode::BAD_REQUEST,
+            UniversalInboxError::ItemNotFound { .. } => StatusCode::BAD_REQUEST,
             UniversalInboxError::AlreadyExists { .. } => StatusCode::CONFLICT,
             UniversalInboxError::Recoverable(_) => StatusCode::INTERNAL_SERVER_ERROR,
             UniversalInboxError::Unexpected(_) => StatusCode::INTERNAL_SERVER_ERROR,