+++ CHANGELOG.md
@@ -35,6 +35,7 @@
 - Handle Slack blocks in attachments
 - Add `default_due_at` setting while syncing Linear assigned issues
 - Add cache directive to task projects search endpoint
+- Create new Todoist sink task if deleted
 
 ## [Initial Version] - 2024-01-27